replace show hide buttons with islands. remove old show hide page level island
